[ALV #4937 #5168] partnerjob: Add UBS mode
XSD are now namespaced (ubs and prospective), and XML readers and
converters are distinct for each data provider. The new UBS mode is very
simple, and not intended for the momement to deal with more than one XML
file.

Note that for UBS, the upstream tasks to retrieve the XML document are not
handled by the Spring Batch job.